% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FancyResult-class.R
\name{$,FancyResult-method}
\alias{$,FancyResult-method}
\title{Access FancyResult slots with \code{$}}
\usage{
\S4method{$}{FancyResult}(x, name)
}
\arguments{
\item{x}{A \linkS4class{FancyResult} object.}

\item{name}{Slot name to extract.}
}
\value{
The contents of the named slot.
}
\description{
Convenience accessor so that \code{result$edges}, \code{result$all_edges},
\code{result$params}, etc. retrieve the corresponding slot of a
\linkS4class{FancyResult} object.
}
