% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FancyResult-class.R
\name{show,FancyResult-method}
\alias{show,FancyResult-method}
\title{Display a FancyResult}
\usage{
\S4method{show}{FancyResult}(object)
}
\arguments{
\item{object}{A \linkS4class{FancyResult} object.}
}
\value{
\code{object}, invisibly.
}
\description{
Compact summary method for \linkS4class{FancyResult} objects.
}
